redo CoreData + refactroing code + fix bug 3

+++ Reciplease/Reciplease/Model/Network/RecipiesJSON.swift
@@ -66,7 +66,7 @@ struct Recipe: Codable {
     let label: String
     let image: String
     let url: String
-    let yield: Int
+    let yield: Double
     let ingredientLines: [String]
     let totalTime: Double
 }
